requirements.txt
ipykernel
dotenv
google-genai
pydantic
tenacity
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests